Decompile unk_02031AF0.s
Something to do with unk save data 40 and 29, not much there, and I chose this smaller file since this was my first day doing any sort of decompiling. Not sure if I did it right, but it passes the sha1 test and compiles properly so I guess it's prolly fine.
I'll try and get better and contribute more in the future when I can =)