Define FLAG_GOT_TM07_FROM_PRYCE
This flag is notable because the game never checks its value, meaning it is functionally useless. It is set when the player gets TM07 from Pryce, but if the player cannot obtain the TM07, no code exists to check the flag and allow the player to obtain the TM07 when they make space in their bag. The only situation where this could occur is when the player has 99 TM07, so this mistake appears to have gone undetected. This mistake exists only in the code for TM07; all other NPCs correctly check the corresponding flag for being unset.